+++ entities/weatheritem.go
@@ -4,6 +4,6 @@ import "gorm.io/gorm"
 
 type WeatherItem struct {
 	gorm.Model
-	Time float32     `json:"time"`
+	Time float64     `json:"time"`
 	Data WeatherData `json:"data" gorm:"foreignKey:ID"`
 }
